Refactor: Medialoader.py
- medialoader 이동에 따른 참조 변화
